# CityPharmacyChain.Api.Client/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CityPharmacyChain.Api.Host",
    "version": "1.0"
  },
  "paths": {
...
    "/Pharmacy/{id}": {
      "get": {
        "tags": [
          "Pharmacy"
        ],
        "summary": "GET запрос по получению объекта класса аптека из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор аптеки",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pharmacy"
        ],
        "summary": "PUT запрос по модификации объекта класса аптека из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор аптеки",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Объект класса аптека в формате JSON",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PharmacyDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PharmacyDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PharmacyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyFullDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyFullDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyFullDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pharmacy"
        ],
        "summary": "DELETE запрос по удалению существующего объекта класса аптека из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор аптеки",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/PharmacyProduct": {
      "get": {
        "tags": [
          "PharmacyProduct"
        ],
        "summary": "GET запрос по получению всех объектов класса связь препарат-аптека из базы данных",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PharmacyProductFullDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PharmacyProductFullDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PharmacyProductFullDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "PharmacyProduct"
        ],
        "summary": "POST запрос по добавлению объекта класса связь препарат-аптека в базу данных",
        "requestBody": {
          "description": "Объект класса связь препарат-аптека в формате JSON",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PharmacyProductDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PharmacyProductDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PharmacyProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyProductFullDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyProductFullDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyProductFullDto"
                }
              }
            }
          }
        }
      }
    },
    "/PharmacyProduct/{id}": {
      "get": {
        "tags": [
          "PharmacyProduct"
        ],
        "summary": "GET запрос по получению объекта класса связь препарат-аптека из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор связи препарат-аптека",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyProductDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyProductDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyProductDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PharmacyProduct"
        ],
        "summary": "PUT запрос по модификации объекта класса связь препарат-аптека из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор связи препарат-аптека",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Объект класса связь препарат-аптека в формате JSON",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PharmacyProductDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PharmacyProductDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PharmacyProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyProductFullDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyProductFullDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyProductFullDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PharmacyProduct"
        ],
        "summary": "DELETE запрос по удалению существующего объекта класса связь препарат-аптека из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор связи препарат-аптека",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/PharmacyProduct/GetTopFivePharmaciesBySoldVolume": {
      "get": {
        "tags": [
          "PharmacyProduct"
        ],
        "summary": "GET запрос по получению коллекции объектов в формате JSON с информацией о топ 5 аптеках по количеству и объёму продаж заданного препарата в заданном временном интервале",
        "parameters": [
          {
            "name": "productName",
            "in": "query",
            "description": "Название препарата",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "Начало временного интервала",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end",
            "in": "query",
            "description": "Конец временного интервала",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductSoldVolumeDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductSoldVolumeDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductSoldVolumeDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/Product/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "GET запрос по получению объекта класса препарат из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор препарата",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "summary": "PUT запрос по модификации объекта класса препарат из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор препарата",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Объект класса препарат в формате JSON",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductFullDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductFullDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductFullDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "summary": "DELETE запрос по удалению существующего объекта класса препарат из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор препарата",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PharmacyDto": {
        "type": "object",
        "properties": {
          "pharmacyNumber": {
            "maximum": 10000,
            "minimum": 0,
            "type": "integer",
            "description": "Номер аптеки",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Название аптеки",
            "nullable": true
          },
          "phoneNumber": {
            "pattern": "^(?!0+$)(\\+\\d{1,3}[- ]?)?(?!0+$)\\d{10,15}$",
            "type": "integer",
            "description": "Телефон аптеки",
            "format": "int64",
            "nullable": true
          },
          "address": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "description": "Адрес аптеки",
            "nullable": true
          },
          "directorFullName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Полное имя директора аптеки",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Класс DTO для сущности класса аптека"
      },
      "PharmacyFullDto": {
        "type": "object",
        "properties": {
          "pharmacyId": {
            "type": "integer",
            "description": "Идентификатор аптеки",
            "format": "int32"
          },
          "pharmacyNumber": {
            "maximum": 10000,
            "minimum": 0,
            "type": "integer",
            "description": "Номер аптеки",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Название аптеки",
            "nullable": true
          },
          "phoneNumber": {
            "pattern": "^(?!0+$)(\\+\\d{1,3}[- ]?)?(?!0+$)\\d{10,15}$",
            "type": "integer",
            "description": "Телефон аптеки",
            "format": "int64",
            "nullable": true
          },
          "address": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "description": "Адрес аптеки",
            "nullable": true
          },
          "directorFullName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Полное имя директора аптеки",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Класс DTO для сущности класса аптека, содержащий её идентификатор"
      },
      "PharmacyProductDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "description": "Идентификатор препарата",
            "format": "int32"
          },
          "pharmacyId": {
            "type": "integer",
            "description": "Идентификатор аптеки",
            "format": "int32"
          },
          "count": {
            "maximum": 1000,
            "minimum": 0,
            "type": "integer",
            "description": "Количество препарата на складе аптеки",
            "format": "int32",
            "nullable": true
          },
          "price": {
            "maximum": 100000,
            "minimum": 0,
            "type": "number",
            "description": "Цена препарата",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Класс DTO для сущности класса связь препарат-аптека"
      },
      "PharmacyProductFullDto": {
        "type": "object",
        "properties": {
          "pharmacyProductId": {
            "type": "integer",
            "description": "Идентификатор связи препарат-аптека",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "description": "Идентификатор препарата",
            "format": "int32"
          },
          "pharmacyId": {
            "type": "integer",
            "description": "Идентификатор аптеки",
            "format": "int32"
          },
          "count": {
            "maximum": 1000,
            "minimum": 0,
            "type": "integer",
            "description": "Количество препарата на складе аптеки",
            "format": "int32",
            "nullable": true
          },
          "price": {
            "maximum": 100000,
            "minimum": 0,
            "type": "number",
            "description": "Цена препарата",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Класс DTO для сущности класса связь препарат-аптека, содержащий её идентификатор"
      },
...
      "ProductDto": {
        "type": "object",
        "properties": {
          "productCode": {
            "maximum": 1000000,
            "minimum": 0,
            "type": "integer",
            "description": "Код препарата",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Название препарата",
            "nullable": true
          },
          "productGroup": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Товарная группа препарата",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Класс DTO для сущности класса препарат"
      },
...
      "ProductFullDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "description": "Идентификатор препарата",
            "format": "int32"
          },
          "productCode": {
            "maximum": 1000000,
            "minimum": 0,
            "type": "integer",
            "description": "Код препарата",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Название препарата",
            "nullable": true
          },
          "productGroup": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Товарная группа препарата",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Класс DTO для сущности класса препарат, содержащий её идентификатор"
      },
      "ProductSoldVolumeDto": {
        "type": "object",
        "properties": {
          "productName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Название препарата",
            "nullable": true
          },
          "pharmacyName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Название аптеки",
            "nullable": true
          },
          "soldCount": {
            "maximum": 1000,
            "minimum": 0,
            "type": "integer",
            "description": "Количество продаж препарата",
            "format": "int32",
            "nullable": true
          },
          "soldVolume": {
            "maximum": 1000000,
            "minimum": 0,
            "type": "integer",
            "description": "Цена препарата, помноженная на количество проданных препаратов этого типа",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Класс DTO для сущности класса количество и объём продаж определённого товара в определённой аптеке"
      }
    }
  }
}